requirements.txt
asgiref==3.3.1
autopep8==1.5.4
beautifulsoup4==4.9.3
cssselect==1.1.0
Django==3.1.5
django-ckeditor==6.0.0
django-js-asset==1.2.2
lxml==4.6.2
markdown2==2.3.10
Pillow==8.1.0
psycopg2-binary==2.9.1
pycodestyle==2.6.0
pyquery==1.4.3
python-decouple==3.4
pytz==2020.5
readtime==1.1.1
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
